+++ mdserver/libvirt.py
@@ -8,6 +8,12 @@
 import xmltodict
 
 from mdserver.database import Database
+from mdserver.util import _removeprefix
+
+
+class LibvirtError(Exception):
+    def __init__(self, message):
+        self.message = message
 
 
 def get_domain_data(domain, net):
@@ -21,6 +27,11 @@ def get_domain_data(domain, net):
     dom = xmltodict.parse(domain)
     ddata["domain_name"] = dom["domain"]["name"]
     ddata["domain_uuid"] = dom["domain"]["uuid"]
+    ddata["domain_metadata"] = {
+        _removeprefix(key, "mdserver:"): dom["domain"]["metadata"][key]
+        for key in dom["domain"]["metadata"]
+        if key.startswith("mdserver:")
+    }
     interfaces = dom["domain"]["devices"]["interface"]
     if not isinstance(interfaces, list):
         interfaces = [interfaces]